Adicionado suporte para contas com numeros decimais

+++ calculadora/src/main/java/com/calculadora/Calculadora.java
@@ -5,7 +5,7 @@ import java.util.Arrays;
 public class Calculadora {
 
     private ArrayList<String> conta = new ArrayList<String>();
-    private int resultado;
+    private double resultado;
 
     public void iniciaCalculo(String contaBruta) {
         contaBruta = "4x+x3x*x7x-x2x+x89x/x3x*x46x-x3";
@@ -21,14 +21,14 @@ public class Calculadora {
     public void calculaMultDiv(int indexListOriginal, int j, String o){
         switch(o){
             case "/":
-                resultado = Integer.parseInt(conta.get(indexListOriginal-1-j)) / Integer.parseInt(conta.get(indexListOriginal+1-j));
+                resultado = Double.parseDouble(conta.get(indexListOriginal-1-j)) / Double.parseDouble(conta.get(indexListOriginal+1-j));
                 conta.set(indexListOriginal-1-j, "" + resultado);
                 conta.remove(indexListOriginal-j);
                 conta.remove(indexListOriginal-j);
             break;
 
             case "*":
-                resultado = Integer.parseInt(conta.get(indexListOriginal-1-j)) * Integer.parseInt(conta.get(indexListOriginal+1-j));
+                resultado = Double.parseDouble(conta.get(indexListOriginal-1-j)) * Double.parseDouble(conta.get(indexListOriginal+1-j));
                 conta.set(indexListOriginal-1-j, "" + resultado);
                 conta.remove(indexListOriginal-j);
                 conta.remove(indexListOriginal-j);
@@ -75,14 +75,14 @@ public class Calculadora {
     //     for (String c : conta) {
     //         switch(c){
     //             case "*":
-    //                 resultado = Integer.parseInt(conta.get(i-1-j)) * Integer.parseInt(conta.get(i+1-j));
+    //                 resultado = Double.parseDouble(conta.get(i-1-j)) * Double.parseDouble(conta.get(i+1-j));
     //                 conta.set(i-1-j, "" + resultado);
     //                 conta.remove(i-j);
     //                 conta.remove(i-j);
     //                 j = j + 2;
     //             break;
     //             case "/":
-    //                 resultado = Integer.parseInt(conta.get(i-1-j)) / Integer.parseInt(conta.get(i+1-j));
+    //                 resultado = Double.parseDouble(conta.get(i-1-j)) / Double.parseDouble(conta.get(i+1-j));
     //                 conta.set(i-1-j, "" + resultado);
     //                 conta.remove(i-j);
     //                 conta.remove(i-j);
@@ -113,7 +113,7 @@ public class Calculadora {
     //     System.out.println(listaMultiplicacao);
     //     int j = 0;
     //     for (Integer m : listaMultiplicacao){
-    //         resultado = Integer.parseInt(conta.get(m-1-j)) * Integer.parseInt(conta.get(m+1-j));
+    //         resultado = Double.parseDouble(conta.get(m-1-j)) * Double.parseDouble(conta.get(m+1-j));
     //         conta.set(m-1-j, "" + resultado);
     //         conta.remove(m-j);
     //         conta.remove(m-j);
@@ -121,7 +121,7 @@ public class Calculadora {
     //     }
 
     //     for (Integer d : listaDivisao){
-    //         resultado = Integer.parseInt(conta.get(d-1-j)) / Integer.parseInt(conta.get(d+1-j));
+    //         resultado = Double.parseDouble(conta.get(d-1-j)) / Double.parseDouble(conta.get(d+1-j));
     //         conta.set(d-1-j, "" + resultado);
     //         conta.remove(d-j);
     //         conta.remove(d-j);
@@ -134,11 +134,11 @@ public class Calculadora {
         for (String c : conta) {
             switch (c) {
                 case "+":
-                    resultado = Integer.parseInt(conta.get(i - 1)) + Integer.parseInt(conta.get(i + 1));
+                    resultado = Double.parseDouble(conta.get(i - 1)) + Double.parseDouble(conta.get(i + 1));
                     break;
 
                 case "-":
-                    resultado = Integer.parseInt(conta.get(i - 1)) - Integer.parseInt(conta.get(i + 1));
+                    resultado = Double.parseDouble(conta.get(i - 1)) - Double.parseDouble(conta.get(i + 1));
                     break;
 
                 default:
